Mild refactoring
